- separate configurator for `kotlinsrc` - separate configurator for `kotlintestsrc` - own `genSources` registrator - use `.toMap()` instead of `.associate { it.first to it.second}`
