+++ modules/java/core/src/sap/commerce/toolset/java/configurator/JavadocsProjectImportConfigurator.kt
@@ -42,7 +42,7 @@ class JavadocsProjectImportConfigurator : ProjectPostImportConfigurator {
             .filter { it.isNonCustomModuleDescriptor }
             .filterNot { it is ConfigModuleDescriptor }
             .forEach { moduleDescriptor ->
-                val module = context.modules[moduleDescriptor.name] ?: run {
+                val module = context.moduleBridges[moduleDescriptor.name] ?: run {
                     thisLogger().warn("Could not find module for ${moduleDescriptor.name}")
                     return@forEach
                 }
