+++ modules/java/core/src/sap/commerce/toolset/java/configurator/contentEntry/util/ContentRootUtil.kt
@@ -34,7 +34,7 @@ import sap.commerce.toolset.util.directoryExists
 import java.nio.file.Path
 import kotlin.io.path.pathString
 
-internal fun ContentRootEntityBuilder.excludeDirectories(
+fun ContentRootEntityBuilder.excludeDirectories(
     context: ProjectImportContext,
     virtualFileUrlManager: VirtualFileUrlManager,
     excludePaths: Collection<Path>
@@ -44,7 +44,7 @@ internal fun ContentRootEntityBuilder.excludeDirectories(
     .map { ExcludeUrlEntity(it, this.entitySource) }
     .let { newExcludedUrls -> this.excludedUrls += newExcludedUrls }
 
-internal fun ContentRootEntityBuilder.addSourceRoots(
+fun ContentRootEntityBuilder.addSourceRoots(
     context: ProjectImportContext,
     virtualFileUrlManager: VirtualFileUrlManager,
     rootEntities: Collection<SourceRootEntityDescriptor>,
@@ -67,23 +67,23 @@ internal fun ContentRootEntityBuilder.addSourceRoots(
     .let { newSourceRoots -> this.sourceRoots += newSourceRoots }
 
 
-internal fun ModuleEntityBuilder.generatedSources(path: Path) = this.sources(
+fun ModuleEntityBuilder.generatedSources(path: Path) = this.sources(
     path = path,
     generated = true
 )
 
-internal fun ModuleEntityBuilder.testGeneratedSources(path: Path) = this.sources(
+fun ModuleEntityBuilder.testGeneratedSources(path: Path) = this.sources(
     sourceRootTypeId = JAVA_TEST_ROOT_ENTITY_TYPE_ID,
     path = path,
     generated = true
 )
 
-internal fun ModuleEntityBuilder.testSources(path: Path) = this.sources(
+fun ModuleEntityBuilder.testSources(path: Path) = this.sources(
     sourceRootTypeId = JAVA_TEST_ROOT_ENTITY_TYPE_ID,
     path = path
 )
 
-internal fun ModuleEntityBuilder.sources(
+fun ModuleEntityBuilder.sources(
     sourceRootTypeId: SourceRootTypeId = JAVA_SOURCE_ROOT_ENTITY_TYPE_ID,
     path: Path,
     generated: Boolean = false,
@@ -95,7 +95,7 @@ internal fun ModuleEntityBuilder.sources(
     javaSourceRoot = { JavaSourceRootPropertiesEntity(generated, packagePrefix, it.entitySource) }
 )
 
-internal fun ModuleEntityBuilder.resources(
+fun ModuleEntityBuilder.resources(
     sourceRootTypeId: SourceRootTypeId = JAVA_RESOURCE_ROOT_ENTITY_TYPE_ID,
     path: Path,
     generated: Boolean = false,
