+++ modules/java/core/src/sap/commerce/toolset/java/configurator/library/CompileModuleLibraryConfigurator.kt
@@ -22,13 +22,14 @@ import com.intellij.platform.workspace.storage.url.VirtualFileUrlManager
 import sap.commerce.toolset.java.JavaConstants
 import sap.commerce.toolset.java.configurator.library.util.*
 import sap.commerce.toolset.project.ProjectConstants
+import sap.commerce.toolset.project.configurator.ModuleLibraryConfigurator
 import sap.commerce.toolset.project.context.ProjectImportContext
 import sap.commerce.toolset.project.context.ProjectModuleConfigurationContext
 import sap.commerce.toolset.project.descriptor.ModuleDescriptor
 import sap.commerce.toolset.project.descriptor.YModuleDescriptor
+import sap.commerce.toolset.project.descriptor.ifNonCustomModuleDescriptor
 import sap.commerce.toolset.project.descriptor.impl.YCommonWebSubModuleDescriptor
 import sap.commerce.toolset.project.descriptor.impl.YWebSubModuleDescriptor
-import sap.commerce.toolset.project.descriptor.isNonCustomModuleDescriptor
 import sap.commerce.toolset.util.directoryExists
 
 class CompileModuleLibraryConfigurator : ModuleLibraryConfigurator<YModuleDescriptor> {
@@ -55,7 +56,7 @@ class CompileModuleLibraryConfigurator : ModuleLibraryConfigurator<YModuleDescri
                 is YCommonWebSubModuleDescriptor -> addAll(moduleDescriptor.webRootJars(virtualFileUrlManager))
             }
 
-            if (moduleDescriptor.isNonCustomModuleDescriptor) {
+            moduleDescriptor.ifNonCustomModuleDescriptor {
                 when (moduleDescriptor) {
                     is YWebSubModuleDescriptor -> addAll(moduleDescriptor.webRootClasses(virtualFileUrlManager))
                     is YCommonWebSubModuleDescriptor -> addAll(moduleDescriptor.webRootClasses(virtualFileUrlManager))
@@ -63,6 +64,7 @@ class CompileModuleLibraryConfigurator : ModuleLibraryConfigurator<YModuleDescri
                 if (importContext.settings.importOOTBModulesInReadOnlyMode) {
                     addAll(moduleDescriptor.classes(virtualFileUrlManager))
                     addAll(moduleDescriptor.resources(virtualFileUrlManager))
+                    addAll(moduleDescriptor.genSources(virtualFileUrlManager))
                     addAll(moduleDescriptor.sources(virtualFileUrlManager))
 
                     when (moduleDescriptor) {
@@ -101,6 +103,7 @@ class CompileModuleLibraryConfigurator : ModuleLibraryConfigurator<YModuleDescri
         .flatMap { moduleDescriptor ->
             buildList {
                 addAll(moduleDescriptor.webRootClasses(virtualFileUrlManager))
+                addAll(moduleDescriptor.genSources(virtualFileUrlManager))
                 addAll(moduleDescriptor.sources(virtualFileUrlManager))
             }
         }
