+++ modules/java/core/src/sap/commerce/toolset/java/configurator/library/PlatformModuleLibraryConfigurator.kt
@@ -20,6 +20,7 @@ package sap.commerce.toolset.java.configurator.library
 
 import sap.commerce.toolset.java.JavaConstants
 import sap.commerce.toolset.java.configurator.library.util.linkProjectLibrary
+import sap.commerce.toolset.project.configurator.ModuleLibraryConfigurator
 import sap.commerce.toolset.project.context.ProjectImportContext
 import sap.commerce.toolset.project.context.ProjectModuleConfigurationContext
 import sap.commerce.toolset.project.descriptor.ModuleDescriptor
