+++ modules/kotlin/core/build.gradle.kts
@@ -35,6 +35,7 @@ sourceSets {
 
 dependencies {
     implementation(project(":shared-core"))
+    implementation(project(":java-core"))
     implementation(project(":project-core"))
     implementation(project(":project-extensioninfo"))
     implementation(project(":project-import-core"))
