+++ modules/kotlin/core/src/sap/commerce/toolset/kotlin/configurator/contentEntry/KotlinSourcesContentRootEntryConfigurator.kt
@@ -15,41 +15,44 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-
-package sap.commerce.toolset.java.configurator.contentEntry
+package sap.commerce.toolset.kotlin.configurator.contentEntry
 
 import com.intellij.platform.workspace.jps.entities.ContentRootEntityBuilder
 import sap.commerce.toolset.java.configurator.contentEntry.util.addSourceRoots
-import sap.commerce.toolset.java.configurator.contentEntry.util.resources
+import sap.commerce.toolset.java.configurator.contentEntry.util.sources
+import sap.commerce.toolset.kotlin.configurator.hasKotlinNatureExtension
 import sap.commerce.toolset.project.ProjectConstants
+import sap.commerce.toolset.project.configurator.ModuleContentRootEntryConfigurator
 import sap.commerce.toolset.project.context.ProjectImportContext
 import sap.commerce.toolset.project.context.ProjectModuleConfigurationContext
 import sap.commerce.toolset.project.descriptor.ModuleDescriptor
-import sap.commerce.toolset.project.descriptor.impl.YBackofficeSubModuleDescriptor
 import sap.commerce.toolset.project.descriptor.isCustomModuleDescriptor
 import java.nio.file.Path
 
-class ResourcesContentEntryConfigurator : ModuleContentEntryConfigurator {
+class KotlinSourcesContentRootEntryConfigurator : ModuleContentRootEntryConfigurator {
 
     override val name: String
-        get() = "Resources"
+        get() = "Kotlin Sources"
 
     override fun isApplicable(
         context: ProjectImportContext,
         moduleDescriptor: ModuleDescriptor
-    ) = moduleDescriptor.isCustomModuleDescriptor || context.settings.importOOTBModulesInWriteMode
+    ) = context.hasKotlinNatureExtension
+        && (moduleDescriptor.isCustomModuleDescriptor || context.settings.importOOTBModulesInWriteMode)
 
     override suspend fun configure(
         context: ProjectModuleConfigurationContext<ModuleDescriptor>,
         contentRootEntity: ContentRootEntityBuilder,
         pathsToIgnore: Collection<Path>
     ) {
-        val moduleDescriptor = context.moduleDescriptor
-        val resourcesPath = moduleDescriptor.moduleRootPath.resolve(ProjectConstants.Directory.RESOURCES)
-        val relativeOutputPath = if (moduleDescriptor is YBackofficeSubModuleDescriptor) "cockpitng" else ""
-        val rootEntities = resourcesPath
-            .let { context.moduleEntity.resources(path = it, relativeOutputPath = relativeOutputPath) }
-            .let { listOf(it) }
+        val moduleEntity = context.moduleEntity
+        val moduleRootPath = context.moduleDescriptor.moduleRootPath
+
+        val rootEntities = buildList {
+            moduleRootPath.resolve(ProjectConstants.Directory.KOTLIN_SRC)
+                .let { moduleEntity.sources(path = it) }
+                .let { add(it) }
+        }
 
         contentRootEntity.addSourceRoots(
             context = context.importContext,
