+++ modules/solr/ui/src/sap/commerce/toolset/solr/options/SolrExecProjectSettingsConfigurableProvider.kt
@@ -99,7 +99,7 @@ class SolrExecProjectSettingsConfigurableProvider(private val project: Project) 
             val connectionService = SolrExecConnectionService.getInstance(project)
             val newSettings = connectionsListPanel.data.map { it.immutable() }
 
-            connectionService.save(newSettings.associate { it.first to it.second })
+            connectionService.save(newSettings.toMap())
 
             if (newSettings.isEmpty()) {
                 originalConnections = connectionService.connections.map { it.mutable() }
